codeapidsl/requirements.txt
requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
click>=8.0.0
requests>=2.25.0
pydantic>=1.8.0
pyyaml>=5.4.0
boto3>=1.18.0
grpcio>=1.39.0
grpcio-tools>=1.39.0
transformers>=4.12.0
torch>=1.9.0
numpy>=1.21.0
pandas>=1.3.0